+++ NEWS.md
@@ -1,5 +1,10 @@
 # trelliscopejs 0.1
 
+- Add auto detection of URL cogs when using cog() explicitly (0.1.18)
+- Fix issue with dplyr issue dealing with multiple factor facet columns (0.1.18)
+- Fix issue with preserving custom panel column class when sorting and filtering with dplyr (0.1.17)
+- Fix bug in auto cognostics with non-numeric/character types, fixes #63 (0.1.16)
+- Fix bug in cognostics types not carrying over correctly, fixes #66 (0.1.15)
 - Update internals to work with latest ggplot2 (0.1.12)
 - Add `plotly_cfg` argument to `facet_trelliscope()` (0.1.11)
 - Update trelliscopejs-lib to avoid double slashes in urls (0.1.10)